+++ rebel-rover/src/components/Footer.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+
 function Footer(){
     return (
       <div>
@@ -230,4 +230,5 @@ function Footer(){
       </div>
     );
     
-}export default Footer;
+}
+export default Footer;
