+++ rebel-rover/src/pages/Blog.jsx
@@ -28,6 +28,7 @@ const Blog = () => {
         </div>
         <BlogSidebar />
       </div>
+      <Footer />
     </div>
   );
 };
